3.AgroScann/Backend/requirements.txt
flask
google-generativeai
Pillow
python-dotenv
flask-cors
